move mosaicViewerExtension to extension.py
